chore: se refactoriza proyecto. Se crea estructura de carpetas. Se crean interfaces Se ajustan imports
