+++ src/types/interfaces.ts
@@ -1,0 +1,18 @@
+export interface Message {
+  text: string;
+  sender: 'user' | 'bot';
+}
+
+export interface DrawerProps {
+  isOpen: boolean;
+  isDesktop: boolean;
+  onToggle: () => void;
+}
+
+export interface ChatMessagesProps {
+  messages: Message[];
+}
+
+export interface ChatInputProps {
+  onSendMessage: (message: string) => void;
+}
